Implement strategy 2
1. Pick a random rule $r$
2. Pick a random recurse $rec$ within the rule $r$
3. Shuffle the rules, find the first for which a merge succeeds + term
size below threshold, add merged to the rules
4. If we can also resolve a reference, resolve it.

- Motivation: instead of only merging if it resolves, just merge. This
  solves the "large enough" problem in strategy (1), which was not
scalable.
- Observation: the term size blows up; we should add a size limit.
- Observation: if you cannot resolve a reference, but the merge
  succeeded, at least we have a larger term, so return it.
- Problem: the strategy focusses on the full set of programs, not one
  program in particular. That is, every iteration a random fragment is
made a bit more "complete". It takes a lot of time for a single fragment
to become complete. Moreover, since every merge increases the size of
the set, and the generation is exponential in the size of the set, the
time grows exponentially. I.e. lots of "warm up" time, which increases
as the language size increases (in terms of # of rules).