Shrink StrategoList
If you shrink by generating a new element in the list, you never get
rid of the element. If you shrink by generating a new list, you are
likely to lose the ambiguity, and shrinking stops.

By leaving out one element in the list, you are likely to keep the
ambiguity while making the list smaller.